Split parquet_format.rs into focused sub-modules (#600)
Split the monolithic parquet_format.rs (~1,085 lines) into:
- parquet_format/mod.rs — public API, re-exports
- parquet_format/schema.rs — schema definitions and validation
- parquet_format/writer.rs — Parquet writing and serialisation
- parquet_format/reader.rs — Parquet reading and deserialisation

All public API items re-exported for backward compatibility.
All existing tests pass without modification.
